Struct wasmtime_fiber::Fiber

pub struct Fiber<'a, Resume, Yield, Return> { /* fields omitted */ }

Implementations

impl<'a, Resume, Yield, Return> Fiber<'a, Resume, Yield, Return>

pub fn new(
stack: FiberStack,
func: impl FnOnce(Resume, &Suspend<Resume, Yield, Return>) -> Return + 'a
) -> Result<Self>

Creates a new fiber which will execute func on the given stack.

This function returns a Fiber which, when resumed, will execute func to completion. When desired the func can suspend itself via Fiber::suspend.

pub fn resume(&self, val: Resume) -> Result<Return, Yield>

Resumes execution of this fiber.

This function will transfer execution to the fiber and resume from where it last left off.

Returns true if the fiber finished or false if the fiber was suspended in the middle of execution.

Panics

Panics if the current thread is already executing a fiber or if this fiber has already finished.

Note that if the fiber itself panics during execution then the panic will be propagated to this caller.

pub fn done(&self) -> bool

Returns whether this fiber has finished executing.

pub fn stack(&self) -> &FiberStack

Gets the stack associated with this fiber.